+++ Tests/SwiftLintFrameworkTests/ParserDiagnosticsTests.swift
@@ -1,0 +1,14 @@
+@testable import SwiftLintFramework
+import XCTest
+
+final class ParserDiagnosticsTests: XCTestCase {
+    func testFileWithParserDiagnostics() {
+        parserDiagnosticsDisabledForTests = false
+        XCTAssertNotNil(SwiftLintFile(contents: "importz Foundation").parserDiagnostics)
+    }
+
+    func testFileWithoutParserDiagnostics() {
+        parserDiagnosticsDisabledForTests = false
+        XCTAssertNil(SwiftLintFile(contents: "import Foundation").parserDiagnostics)
+    }
+}